HW4

+++ hw4/problem2/gauss.c
@@ -241,7 +241,7 @@ void gauss() {
 		t2 = MPI_Wtime();
 		tick = MPI_Wtick(); // Time in seconds that an MPI tick represents
 
-		printf("\nElapsed time: %25.3d miliseconds\n", (t2-t1)/(1000*tick));
+		printf("\nElapsed time: %d miliseconds\n", ((double)(t2-t1))/(1000*tick));
 	}
 	
 }
